Rename 'obd_pages' package to 'presentation'
