+++ app/src/main/java/ecb/manifest/kowalski/obd_scan/ui/presentation/BluetoothUiState.kt
@@ -1,0 +1,8 @@
+package ecb.manifest.kowalski.obd_scan.ui.presentation
+
+import ecb.manifest.kowalski.obd_scan.bluetooth.BluetoothDeviceDomain
+
+data class BluetoothUiState(
+    val scannedDevices: List<BluetoothDeviceDomain> = emptyList(),
+    val pairedDevices: List<BluetoothDeviceDomain> = emptyList(),
+)
