+++ app/src/main/java/ecb/manifest/kowalski/obd_scan/ui/presentation/fuel_page/FuelFragment.kt
@@ -1,4 +1,4 @@
-package ecb.manifest.kowalski.obd_scan.ui.obd_pages.status_page
+package ecb.manifest.kowalski.obd_scan.ui.presentation.fuel_page
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -9,15 +9,14 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 
-
 @Preview
 @Composable
-fun StatusPage() {
+fun FuelPage() {
     Column(
         modifier = Modifier.fillMaxSize(),
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
-        Text(text = "Status Page Component Placeholder")
+        Text(text = "Fuel Page Component Placeholder")
     }
 }
